+++ zoobudget.h
@@ -1,24 +1,24 @@
 #ifndef ZOOBUDGET_H
 #define ZOOBUDGET_H
 
-#include "zoomessage.h"
 #include "zooobject.h"
+#include "zoomessage.h"
 
 
 class ZooBudget : public ZooObject
 {
-    int money;
+    double money;
     static ZooBudget *m_budget;
 public:
     ~ZooBudget();
-    int addMoney(int amount);
-    bool removeMoney(int amount);
-    int getMoney();
+    double addMoney(double amount);
+    bool removeMoney(double amount);
+    double getMoney();
     void testMe();
     static ZooBudget *getInstance();
 private:
     ZooBudget(const QString &name);
-    void setMoney(int value);
+    void setMoney(double value);
 };
 
 #endif // ZOOBUDGET_H
